Allow FLAC tags to have duplicate entries (fixes #2)
This means the FLAC file can have two or more ARTIST tags or any other
tag. Easytag will do this when you have artist field as "Artist1 -
Artist2".

With this commit, the FLAC decoder will combine multiple tags, seperated
by " - ". This is the best way to support AAC, as it does not allow for
multiple tag. In the future, the string join could be moved to the AAC
encoder class, but this would require extra work.